+++ R/RcppExports.R
@@ -17,10 +17,13 @@
 #' @param summax maximum number of terms to be considered in the truncated sum.
 #' @useDynLib mpcmp, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
+#' @return A numeric vector giving the (approximate) log of the COM-Poisson
+#' normalizing constant \eqn{\log Z(\lambda, \nu)} for each element of
+#' \code{log_lambda} and \code{nu}.
 #' @references 
 #' Ribeiro Jr, E. E., Zeviani, W. M., Demétrio, C. G. B. (2019) \code{cmpreg}: 
 #' Reparametrized COM-Poisson Regression Models. R package version 0.0.1.
-#' @export
+#' @keywords internal
 #' 
 logZ_c <- function(log_lambda, nu, summax) {
     .Call(`_mpcmp_logZ_c`, log_lambda, nu, summax)