+++ R/comp_lambdas.R
@@ -24,10 +24,10 @@
 #' appropriate lambda values.
 #'
 #' @name comp_lambdas
+#' @keywords internal
 NULL
 
 #' @rdname comp_lambdas
-#' @export
 comp_lambdas <- function(mu, nu, lambdalb = 1e-10, lambdaub = 1000,
                          maxlambdaiter = 1e3, tol = 1e-6, lambdaint = 1, summax = 100) {
   df <- CBIND(mu = mu, nu = nu, lambda = lambdaint, lb = lambdalb, ub = lambdaub)
@@ -84,7 +84,6 @@ comp_lambdas <- function(mu, nu, lambdalb = 1e-10, lambdaub = 1000,
 }
 
 #' @rdname comp_lambdas
-#' @export
 comp_lambdas_fixed_ub <- function(mu, nu, lambdalb = 1e-10, lambdaub = 1000,
                                   maxlambdaiter = 1e3, tol = 1e-6, lambdaint = 1,
                                   summax = 100) {