+++ tests/testthat/test_summarise.R
@@ -1,26 +1,30 @@
-context("Test Summarise and Extract info from CMP object")
 library(mpcmp)
 data("attendance")
 M.attendance <- glm.cmp(daysabs ~ gender + math + prog,
   data = attendance
 )
 newdataframe <- data.frame(
-  gender= "female", math = 70, prog = "Academic")
-pred1 <- predict(M.attendance, 
-                 newdata = newdataframe, 
-                 se.fit = TRUE)
-pred2 <- predict(M.attendance, 
-                 newdata = newdataframe, 
-                 type = "response",
-                 se.fit = TRUE)
+  gender = "female", math = 70, prog = "Academic"
+)
+pred1 <- predict(M.attendance,
+  newdata = newdataframe,
+  se.fit = TRUE
+)
+pred2 <- predict(M.attendance,
+  newdata = newdataframe,
+  type = "response",
+  se.fit = TRUE
+)
 
 M.sit <- glm.cmp(formula = ninsect ~ extract, formula_nu = ~extract, data = sitophilus)
 
-M.sit_summary_coef <- 
+M.sit_summary_coef <-
   c(-1.455, -0.589, -0.572, -0.076)
-names(M.sit_summary_coef) <- 
-  c("(Intercept)", "extractLeaf",
-    "extractBranch", "extractSeed")
+names(M.sit_summary_coef) <-
+  c(
+    "(Intercept)", "extractLeaf",
+    "extractBranch", "extractSeed"
+  )
 test_that("Test residuals", {
   expect_equal(
     -0.264,
@@ -87,16 +91,14 @@ test_that("Test sumamry", {
     capture_output_lines(print(summary(M.attendance)))[24],
     "AIC: 1739.026 "
   )
-  expect_is(summary(M.attendance)$coefficients, "matrix")
+  expect_true(is.matrix(summary(M.attendance)$coefficients))
   expect_vector(as.vector(summary(M.attendance)$coefficients),
     ptype = numeric(), size = 20
   )
-  expect_is(summary(M.sit)$coefficients, "matrix")
-  expect_is(summary(M.sit)$coef.table_beta, "matrix")
-  expect_is(summary(M.sit)$coef.table_gamma, "matrix")
-  expect_equal(
-    round(summary(M.sit)$coef.table_gamma[,3], 3),
-    M.sit_summary_coef)
+  expect_true(is.matrix(summary(M.sit)$coefficients))
+  expect_true(is.matrix(summary(M.sit)$coef.table_beta))
+  expect_true(is.matrix(summary(M.sit)$coef.table_gamma))
+  expect_snapshot(summary(M.sit)$coef.table_gamma)
 })
 
 test_that("Test rstandard", {
@@ -112,7 +114,7 @@ test_that("Test rstandard", {
 
 test_that("Test influence", {
   infl <- influence.cmp(M.attendance)
-  expect_is(infl, class = "list")
+  expect_type(infl, "list")
   expect_equal(unname(infl$h[1]), 0.01152283)
   expect_equal(unname(round(infl$dev_res[1], 4)), -0.2644)
   expect_equal(unname(round(infl$pear_res[1], 4)), -0.2437)
@@ -124,13 +126,13 @@ test_that("Test hatvalues", {
 })
 
 test_that("Test cooks.distance", {
-  expect_equal(unname(cooks.distance.cmp(M.attendance)[1]), 0.0001400528)
+  expect_snapshot(cooks.distance.cmp(M.attendance))
   expect_length(cooks.distance.cmp(M.attendance), 314)
 })
 
 test_that("Test vcov", {
-  expect_is(vcov(M.attendance), class = "matrix")
-  expect_is(vcov(M.sit), class = "list")
+  expect_true(is.matrix(vcov(M.attendance)))
+  expect_type(vcov(M.sit), "list")
 })
 
 test_that("Test broom", {
@@ -149,30 +151,40 @@ test_that("Test broom", {
     )
   )
   expect_length(glance(M.attendance), 8)
-  expect_is(glance(M.attendance), class = "tbl_df")
+  expect_s3_class(glance(M.attendance), "tbl_df")
   expect_length(glance(M.sit), 8)
-  expect_is(glance(M.sit), class = "tbl_df")
-  expect_is(augment(M.attendance), class = "tbl_df")
+  expect_s3_class(glance(M.sit), "tbl_df")
+  expect_s3_class(augment(M.attendance), "tbl_df")
   expect_length(augment(M.attendance), 9)
   expect_length(augment(M.sit), 9)
 })
 
 
-
 test_that("Test patching predict with new data", {
-  expect_equal( 
-    round(c(predict(M.attendance, newdata = newdataframe)),
-          5),
-    round(1.846711,5))
-  expect_equal(round(c(predict(M.attendance, 
-                  newdata = newdataframe, 
-                  type = "response")),
-        5), round(6.33894,5))
-  expect_length(predict(M.attendance, 
-                          newdata = newdataframe, 
-                          se.fit=TRUE), 2)
-  expect_equal(round(as.numeric(pred1$se.fit), 5), 
-               round(0.1172887,5))
-  expect_equal(round(as.numeric(pred2$se.fit), 5),
-               round(0.7434861,5))
+  expect_equal(
+    round(
+      c(predict(M.attendance, newdata = newdataframe)),
+      5
+    ),
+    round(1.846711, 5)
+  )
+  expect_equal(round(
+    c(predict(M.attendance,
+      newdata = newdataframe,
+      type = "response"
+    )),
+    5
+  ), round(6.33894, 5))
+  expect_length(predict(M.attendance,
+    newdata = newdataframe,
+    se.fit = TRUE
+  ), 2)
+  expect_equal(
+    round(as.numeric(pred1$se.fit), 5),
+    round(0.1172887, 5)
+  )
+  expect_equal(
+    round(as.numeric(pred2$se.fit), 5),
+    round(0.7434861, 5)
+  )
 })